+++ src/roiextractors/extractors/miniscopeimagingextractor/miniscopeimagingextractor.py
@@ -243,3 +243,9 @@ class _MiniscopeSingleVideoExtractor(ImagingExtractor):
             )
 
         return self.get_series(start_sample=start_frame, end_sample=end_frame)
+
+    def get_native_timestamps(
+        self, start_sample: Optional[int] = None, end_sample: Optional[int] = None
+    ) -> Optional[np.ndarray]:
+        # Miniscope videos do not have native timestamps
+        return None